:wrench: rename everything to bearer_cli
