+++ unittests/scans/bearer_cli/bearer_cli_many_vul.json
