+++ unittests/scans/bearer_cli/bearer_cli_one_vul.json
