+++ unittests/tools/test_bearer_cli_parser.py
@@ -1,12 +1,12 @@
 from django.test import TestCase
-from dojo.tools.bearer.parser import BearerParser
+from dojo.tools.bearer_cli.parser import BearerParser
 from dojo.models import Test
 
 
 class TestBearerParser(TestCase):
 
     def test_bearer_parser_with_one_vuln_has_one_findings(self):
-        testfile = open("unittests/scans/bearer/bearer_one_vul.json")
+        testfile = open("unittests/scans/bearer_cli/bearer_cli_one_vul.json")
         parser = BearerParser()
         findings = parser.get_findings(testfile, Test())
         testfile.close()
@@ -20,7 +20,7 @@ class TestBearerParser(TestCase):
         self.assertEqual(581, findings[0].line)
 
     def test_bearer_parser_with_many_vuln_has_many_findings(self):
-        testfile = open("unittests/scans/bearer/bearer_many_vul.json")
+        testfile = open("unittests/scans/bearer_cli/beare_cli_many_vul.json")
         parser = BearerParser()
         findings = parser.get_findings(testfile, Test())
         testfile.close()
